build the image without user and use setpriv
The issue to solve is, that CI may run as unprivileged user, as it does
in GitHub Actions, and we don't want to write files as root that the
user can't access.

With this change, we don't set the UID on build time anymore, but rather
check the ownership of a .git directory or, if that doesn't exist, the
working directory. This makes the image more portable and the custom
root image obsolete.