test: rename local branch
using 'test' collides with test/data1 and test/data2.
This is only an issue when pushing those branches, but it's annoying.

I'm not really sure why I renamed the branch in the first place,
but it's probably a syntax quirk having to write master:<target> when
combining the fetch with 'tag <tag>' syntax.